Prohibition of access by others

+++ app/Http/Controllers/TasksController.php
@@ -7,6 +7,8 @@ use App\Task;
 
 class TasksController extends Controller
 {
+    
+    // タスク一覧
     public function index()
     {
         $data = [];
@@ -21,6 +23,7 @@ class TasksController extends Controller
         return view('tasks.index', $data);
     }
     
+    // タスク投稿コントローラ
     public function store(Request $request)
     {
         $request->validate(['content' => 'required|max:255', 'status' => 'required|max:10',]);
@@ -31,22 +34,15 @@ class TasksController extends Controller
         return redirect('/');
     }
 
-
-
-
-
     public function create()
     {
         $task = new Task;
         return view('tasks.create', ['task' => $task,]);
     }
 
- 
-
-
-
     public function show($id)
     {
+
         $task = Task::findOrFail($id);
         return view('tasks.show', ['task' => $task,]);
     }
@@ -63,16 +59,22 @@ class TasksController extends Controller
     {
         $request->validate(['content' => 'required|max:255', 'status' => 'required|max:10',]);
         $task = Task::findOrFail($id);
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-        return redirect('/');
+        // ログインユーザーのidが一致した場合のみという条件
+        if (\Auth::id() === $task->user_id) {    
+            $task->content = $request->content;
+            $task->status = $request->status;
+            $task->save();
+            return redirect('/');
+        }    
     }
 
     public function destroy($id)
     {
         $task = Task::findOrFail($id);
-        $task->delete();
-        return redirect('/');
+        // ログインユーザーのidが一致した場合のみという条件
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return redirect('/');
+        }
     }
 }
